runtime: move get_contract_code logic to RuntimeExt
There's exactly one use of this function and as a future imporvement
this could be called by the contract runtime directly in order to obtain
the contract code to execute. Transaction runtime then doesn't need to
worry about any of the caching logic specifics internal to the contract
runtime.
