+++ runtime/runtime/src/actions.rs
@@ -44,28 +44,6 @@ use near_vm_runner::precompile_contract;
 use near_vm_runner::ContractCode;
 use near_wallet_contract::{wallet_contract, wallet_contract_magic_bytes};
 
-use std::sync::Arc;
-
-/// Returns `ContractCode` (if exists) for the given `account` or returns `StorageError`.
-/// For ETH-implicit accounts returns `Wallet Contract` implementation that it is a part
-/// of the protocol and it's cached in memory.
-fn get_contract_code(
-    runtime_ext: &RuntimeExt,
-    account: &Account,
-    protocol_version: ProtocolVersion,
-) -> Result<Option<Arc<ContractCode>>, StorageError> {
-    let account_id = runtime_ext.account_id();
-    let code_hash = account.code_hash();
-    if checked_feature!("stable", EthImplicitAccounts, protocol_version)
-        && account_id.get_account_type() == AccountType::EthImplicitAccount
-    {
-        let chain_id = runtime_ext.chain_id();
-        assert!(&code_hash == wallet_contract_magic_bytes(&chain_id).hash());
-        return Ok(Some(wallet_contract(&chain_id)));
-    }
-    Ok(runtime_ext.get_code(code_hash).map(Arc::new))
-}
-
 /// Runs given function call with given context / apply state.
 pub(crate) fn execute_function_call(
     apply_state: &ApplyState,
@@ -140,14 +118,7 @@ pub(crate) fn execute_function_call(
         Err(VMRunnerError::CacheError(CacheError::ReadError(err)))
             if err.kind() == std::io::ErrorKind::NotFound =>
         {
-            if checked_feature!("stable", ChunkNodesCache, protocol_version) {
-                runtime_ext.set_trie_cache_mode(TrieCacheMode::CachingShard);
-            }
-            let code = match get_contract_code(
-                &runtime_ext,
-                account,
-                apply_state.current_protocol_version,
-            ) {
+            let code = match runtime_ext.get_contract(account) {
                 Ok(Some(code)) => code,
                 Ok(None) => {
                     let error =
@@ -160,9 +131,6 @@ pub(crate) fn execute_function_call(
                     return Err(RuntimeError::StorageError(e));
                 }
             };
-            if checked_feature!("stable", ChunkNodesCache, protocol_version) {
-                runtime_ext.set_trie_cache_mode(TrieCacheMode::CachingChunk);
-            }
             let r = near_vm_runner::run(
                 account,
                 Some(&code),
